requirements.txt
asgiref==3.2.3
Django==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
Faker==4.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
